Logic changes.

+++ appSpring/src/main/java/appSpring/restController/UserRestController.java
@@ -2,9 +2,13 @@
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,7 +17,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.fasterxml.jackson.annotation.JsonView;
 
-import appSpring.component.UserComponent;
 import appSpring.model.Action;
 import appSpring.model.Fine;
 import appSpring.model.User;
@@ -33,8 +36,6 @@ public class UserRestController {
 	private ActionService actionService;
 	@Autowired
 	private LogicService logicService;
-	@Autowired
-	private UserComponent userComponent;
 
 	@JsonView(UserDetail.class)
 	@RequestMapping(value = "/", method = RequestMethod.POST)
@@ -50,8 +51,9 @@ public class UserRestController {
 
 	@JsonView(UserDetail.class)
 	@RequestMapping(value = "/all", method = RequestMethod.GET)
-	public ResponseEntity<List<User>> getUsers() {
+	public ResponseEntity<List<User>> getUsers(HttpSession session) {
 
+		session.setMaxInactiveInterval(-1);
 		List<User> users = userService.findAll();
 		if (users != null) {
 			return new ResponseEntity<>(users, HttpStatus.OK);
@@ -62,11 +64,16 @@ public class UserRestController {
 
 	@JsonView(UserDetail.class)
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public ResponseEntity<User> getUser(@PathVariable int id) {
+	public ResponseEntity<User> getUser(@PathVariable int id, Authentication authentication, HttpSession session,
+			HttpServletRequest request) {
 
+		session.setMaxInactiveInterval(-1);
 		User user = userService.findOne(id);
 		if (user != null) {
-			return new ResponseEntity<>(user, HttpStatus.OK);
+			if ((authentication.getName().contains(user.getName())) || (request.isUserInRole("ADMIN")))
+				return new ResponseEntity<>(user, HttpStatus.OK);
+			else
+				return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 		} else {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
@@ -74,14 +81,19 @@ public class UserRestController {
 
 	@JsonView(UserDetail.class)
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<User> deleteUser(@PathVariable Integer id) {
+	public ResponseEntity<User> deleteUser(@PathVariable Integer id, HttpSession session,
+			Authentication authentication) {
 
+		session.setMaxInactiveInterval(-1);
 		User userSelected = userService.findOne(id);
 		if (userSelected != null) {
+			if (authentication.getName().contains(userSelected.getName())) {
+				return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+			}
 			List<Action> actions = actionService.findAll();
 			for (Action action : actions) {
 				if ((action.getDateLoanReturn() == null) && (action.getUser() == userSelected)) {
-					return new ResponseEntity<>(HttpStatus.CONFLICT);
+					return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
 				}
 			}
 			userService.delete(userSelected);
@@ -93,15 +105,17 @@ public class UserRestController {
 
 	@JsonView(UserDetail.class)
 	@RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-	public ResponseEntity<User> putUser(@PathVariable Integer id, @RequestBody User userUpdated) {
+	public ResponseEntity<User> putUser(@PathVariable Integer id, @RequestBody User userUpdated,
+			HttpSession session, HttpServletRequest request, Authentication authentication) {
 
+		session.setMaxInactiveInterval(-1);
 		User user = userService.findOne(id);
 		if ((user != null) && (user.getId() == userUpdated.getId())) {
-			if (userComponent.getLoggedUser() == user) {
+			if ((authentication.getName().contains(user.getName())) || (request.isUserInRole("ADMIN"))) {
 				userService.save(userUpdated);
 				return new ResponseEntity<>(userUpdated, HttpStatus.OK);
 			} else {
-				return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+				return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 			}
 		} else {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
